update the sample and move the common stuff into the new generic file, add one more test
