+++ dflag/dyngeneric.go
@@ -13,6 +13,37 @@ import (
 	"golang.org/x/exp/constraints"
 )
 
+// DynamicFlagValue interface is a tag to know if a type is dynamic or not.
+type DynamicFlagValue interface {
+	IsDynamicFlag() bool
+}
+
+// DynamicJSONFlagValue is a tag interface for JSON dynamic flags.
+type DynamicJSONFlagValue interface {
+	IsJSON() bool
+}
+
+// DynamicFlagValueTag is a struct all dynamic flag inherit for marking they are dynamic.
+type DynamicFlagValueTag struct{}
+
+// IsDynamicFlag always returns true.
+func (*DynamicFlagValueTag) IsDynamicFlag() bool {
+	return true
+}
+
+// A flag is dynamic if it implements DynamicFlagValue (which all the dyn* do)
+
+// IsFlagDynamic returns whether the given Flag has been created in a Dynamic mode.
+func IsFlagDynamic(f *flag.Flag) bool {
+	df, ok := f.Value.(DynamicFlagValue)
+	if !ok {
+		return false
+	}
+	return df.IsDynamicFlag() // will clearly return true if it exists
+}
+
+// ---- Generics section ---
+
 type DynValueTypes interface {
 	bool | time.Duration | float64 | int64 | string
 }
