+++ dflag/dyngeneric_test.go
@@ -4,6 +4,7 @@
 package dflag
 
 import (
+	"flag"
 	"testing"
 )
 
@@ -13,4 +14,12 @@ func TestParse_BadType(t *testing.T) {
 	_, err := Parse[uint8]("23")
 	assert.Error(t, err, "Expecting unpected type error")
 	assert.Equal(t, err.Error(), "unexpected type uint8", "message/error should match")
+}
+
+func TestDflag_NonDynamic(t *testing.T) {
+	set := flag.NewFlagSet("foobar", flag.ContinueOnError)
+	set.Bool("notdyn", false, "...")
+	static := set.Lookup("notdyn")
+	assert.True(t, static != nil)
+	assert.False(t, IsFlagDynamic(static))
 }
